refactor: update imports for HorizontalScrollSection and enhance loading components for better UI consistency
